requirements.txt
osxphotos>=0.68.0
pytumblr>=0.1.0
anthropic>=0.25.0
Pillow>=10.3.0
python-dotenv>=1.0.1
timezonefinder>=6.5.0
